+++ lib/pages/denue.dart
@@ -1,7 +1,6 @@
-import 'package:api_denue/widgets/controller_map.dart';
-import 'package:google_maps_flutter/google_maps_flutter.dart';
+import 'package:api_denue/services/apiDenue.dart';
+import 'package:api_denue/widgets/recetaSeleccionada.dart';
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
 
 class DenueInegi extends StatefulWidget {
   const DenueInegi({super.key});
@@ -16,164 +15,121 @@ class _DenueInegiState extends State<DenueInegi> {
     super.initState();
   }
 
-  var listaEconomia = ["Restaurantes", "Hoteles", "Camiones", "Pesca"];
-  var listaDistancia = ["300", "500", "1000"];
-  String seccionDistancia = "Distancia";
-  String seccionEconomia = "Economia";
+  var lista = ["Economia1", "Economia2", "Economia3"];
+  String seccion = "Selec. Economia";
   final codPostal = TextEditingController();
   int noPostal = 0;
-
-  final CameraPosition _kGooglePlex = const CameraPosition(
-    target: LatLng(16.755841143621865, -93.12323672443239),
-    zoom: 14.4746,
-  );
-
-  late LatLng latitud = const LatLng(0, 0);
-  bool botonBuscar = false;
-  bool botonEconomia = false;
-  bool botonDistancia = false;
-  bool clickUbicacion = false;
 
   @override
   Widget build(BuildContext context) {
     final Size size = MediaQuery.of(context).size;
 
-    return ChangeNotifierProvider<ControllerMap>(
-      create: (_) => ControllerMap(),
-      child: Scaffold(
-        appBar: AppBar(
-          title: const Text('Denue'),
-        ),
-        body: Stack(
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Denue'),
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: () {
+          Navigator.pushReplacementNamed(context, 'listDenue', arguments: {
+            'data': [
+              'comercio',
+              '16.2320885648689',
+              '-93.90821390642238',
+              '250'
+            ]
+          });
+        },
+        child: const Icon(Icons.info_outline),
+      ),
+      body: Container(
+        width: size.width,
+        height: size.height,
+        color: null,
+        child: Column(
           children: [
-            SizedBox(
+            Container(
+              color: Colors.transparent,
               width: size.width,
-              height: size.height,
-              child: Column(
+              height: size.height * 0.1,
+              child: Row(
+                mainAxisSize: MainAxisSize.max,
+                mainAxisAlignment: MainAxisAlignment.spaceAround,
                 children: [
-                  SizedBox(
-                    width: size.width,
-                    height: size.height * 0.1,
-                    child: Row(
-                      mainAxisSize: MainAxisSize.max,
-                      mainAxisAlignment: MainAxisAlignment.spaceAround,
-                      children: [
-                        Container(
-                          // color: Colors.pink,
-                          width: size.width * 0.4,
-                          padding: const EdgeInsets.all(5),
-                          child: DropdownButton(
-                            items: listaEconomia.map((String resultado) {
-                              return DropdownMenuItem(
-                                  value: resultado, child: Text(resultado));
-                            }).toList(),
-                            onChanged: ((value) {
-                              setState(() {
-                                seccionEconomia = value.toString();
-                                botonEconomia = true;
-                                if (clickUbicacion == true &&
-                                    botonDistancia == true &&
-                                    botonEconomia == true) {
-                                  botonBuscar = true;
-                                }
-                              });
-                            }),
-                            hint: Text(seccionEconomia),
-                            borderRadius: BorderRadius.circular(10),
-                          ),
-                        ),
-                        Container(
-                          width: size.width * 0.4,
-                          padding: const EdgeInsets.all(5),
-                          child: DropdownButton(
-                            items: listaDistancia.map((String resultado) {
-                              return DropdownMenuItem(
-                                  value: resultado, child: Text(resultado));
-                            }).toList(),
-                            onChanged: ((value) {
-                              setState(() {
-                                seccionDistancia = value.toString();
-                                botonDistancia = true;
-                                if (clickUbicacion == true &&
-                                    botonDistancia == true &&
-                                    botonEconomia == true) {
-                                  botonBuscar = true;
-                                }
-                              });
-                            }),
-                            hint: Text(seccionDistancia),
-                            borderRadius: BorderRadius.circular(10),
-                          ),
-                        ),
-                      ],
+                  Container(
+                    // color: Colors.blue,
+                    width: size.width * 0.3,
+                    padding: const EdgeInsets.all(5),
+                    child: TextField(
+                      controller: codPostal,
+                      autofocus: false,
+                      maxLength: 5,
+                      keyboardType: TextInputType.number,
+                      decoration: const InputDecoration(
+                        hintText: 'Cod. Postal',
+                        // enabledBorder: OutlineInputBorder(
+                        //     borderSide: BorderSide(color: Colors.black)),
+                        // focusedBorder: OutlineInputBorder(
+                        //     borderSide: BorderSide(color: Colors.black))
+                      ),
                     ),
                   ),
+                  Container(
+                    // color: Colors.pink,
+                    width: size.width * 0.4,
+                    padding: const EdgeInsets.all(5),
+                    child: DropdownButton(
+                      items: lista.map((String resultado) {
+                        return DropdownMenuItem(
+                            value: resultado, child: Text(resultado));
+                      }).toList(),
+                      onChanged: ((value) {
+                        setState(() {
+                          seccion = value.toString();
+                        });
+                      }),
+                      hint: Text(seccion),
+                      borderRadius: BorderRadius.circular(10),
+                    ),
+                  ),
+                  Container(
+                      color: Colors.transparent,
+                      width: size.width * 0.2,
+                      // padding: const EdgeInsets.all(5),
+                      child: ElevatedButton(
+                          onPressed: () {
+                            noPostal = int.parse(codPostal.text);
+                            showDialog(
+                                context: context,
+                                barrierDismissible: false,
+                                builder: (context) => AlertDialog(
+                                      title: Text(
+                                          'Codigo postal: $noPostal y Seleccion de economia: $seccion'),
+                                      actions: [
+                                        TextButton(
+                                            onPressed: () {
+                                              Navigator.pop(context);
+                                            },
+                                            child: const Text('OK'))
+                                      ],
+                                    ));
+                            print(noPostal);
+                          },
+                          child: const Text('Buscar')))
                 ],
               ),
             ),
-            Container(
-              padding: EdgeInsets.only(top: size.height * 0.1),
-              width: size.width,
-              height: size.height * 0.7,
-              child: Consumer<ControllerMap>(
-                builder: (_, controller, __) => GoogleMap(
-                  myLocationEnabled: true,
-                  mapType: MapType.normal,
-                  initialCameraPosition: _kGooglePlex,
-                  onMapCreated: controller.onMapCreated,
-                  markers: controller.markers,
-                  onTap: (argument) {
-                    latitud = controller.onTap(argument);
-                    clickUbicacion = true;
-                    setState(() {});
-                    if (clickUbicacion == true &&
-                        botonDistancia == true &&
-                        botonEconomia == true) {
-                      botonBuscar = true;
-                    }
+            SizedBox(
+              height: size.height * 0.5,
+              child: ListView.builder(
+                itemCount: 5,
+                itemBuilder: (_, index) => InkWell(
+                  child: const RecetaSeleccionada(
+                      prueba1: "colorCaducidad", prueba2: "Color asda"),
+                  onTap: () {
+                    print("Se esta tocando");
                   },
                 ),
               ),
-            ),
-            Center(
-              child: Container(
-                  // color: Colors.pink,
-                  margin: EdgeInsets.only(top: size.height * 0.6),
-                  height: size.height * 0.05,
-                  width: size.width * 0.4,
-                  // padding: const EdgeInsets.all(5),
-                  child: ElevatedButton(
-                      onPressed: botonBuscar
-                          ? () {
-                              print("Economia: " + seccionEconomia);
-                              print("Distancia: " + seccionDistancia);
-                              print("Latitud: " +
-                                  latitud.latitude.toString() +
-                                  " " +
-                                  latitud.longitude.toString());
-                              print("Latitud: " + latitud.toString());
-                              if (seccionEconomia != "Economia" &&
-                                  seccionDistancia != "Distancia" &&
-                                  latitud.latitude.toString() != "0.0" &&
-                                  latitud.longitude.toString() != "0.0") {
-                                print("Es correcto");
-                                Navigator.pushReplacementNamed(
-                                    context, 'listDenue',
-                                    arguments: {
-                                      'data': [
-                                        seccionEconomia,
-                                        latitud.latitude.toString(),
-                                        latitud.longitude.toString(),
-                                        seccionDistancia
-                                      ]
-                                    });
-                              } else {
-                                print("No has elegido uno");
-                              }
-                              //
-                            }
-                          : null,
-                      child: const Text('Buscar'))),
             )
           ],
         ),
