+++ pubspec.yaml
@@ -35,8 +35,6 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   http: ^0.13.5
-  google_maps_flutter: ^2.2.0
-  provider: ^6.0.3
 
 dev_dependencies:
   flutter_test:
